refactor: modularize api/submit-lead.ts and network utilities
- Created `lib/network.ts` for Edge Function network logic.
- Created `lib/lead-logic.ts` for lead validation and sanitization.
- Created `services/hubspot.ts` for CRM integration.
- Refactored `api/submit-lead.ts` and `api/generate.ts` to use these modules.
- Updated `tests/submit-lead.test.ts` imports.
