+++ services/hubspot.ts
@@ -154,7 +154,8 @@ export async function getContactIdByEmail(token: string, email: string): Promise
     }
 
     if (!response.ok) {
-        throw new Error(`CONTACT_SEARCH_FAILED:${response.status}`);
+        const detail = await response.text().catch(() => '');
+        throw new Error(`CONTACT_SEARCH_FAILED:${response.status}:${detail}`);
     }
 
     const data = (await response.json().catch(() => ({}))) as HubSpotSearchResponse;
@@ -177,7 +178,8 @@ export async function updateContactProperties(token: string, contactId: string, 
     }
 
     if (!response.ok) {
-        throw new Error(`CONTACT_UPDATE_FAILED:${response.status}`);
+        const detail = await response.text().catch(() => '');
+        throw new Error(`CONTACT_UPDATE_FAILED:${response.status}:${detail}`);
     }
 }
 
@@ -195,7 +197,8 @@ export async function createDeal(token: string, properties: Record<string, strin
     }
 
     if (!response.ok) {
-        throw new Error(`DEAL_CREATE_FAILED:${response.status}`);
+        const detail = await response.text().catch(() => '');
+        throw new Error(`DEAL_CREATE_FAILED:${response.status}:${detail}`);
     }
 
     const data = (await response.json().catch(() => ({}))) as HubSpotObjectResponse;
@@ -225,7 +228,8 @@ export async function associateDealToContact(token: string, dealId: string, cont
     }
 
     if (!response.ok) {
-        throw new Error(`DEAL_ASSOCIATION_FAILED:${response.status}`);
+        const detail = await response.text().catch(() => '');
+        throw new Error(`DEAL_ASSOCIATION_FAILED:${response.status}:${detail}`);
     }
 }
 
@@ -248,7 +252,8 @@ export async function createFallbackNote(token: string, contactId: string, body:
     }
 
     if (!noteResponse.ok) {
-        throw new Error(`NOTE_CREATE_FAILED:${noteResponse.status}`);
+        const detail = await noteResponse.text().catch(() => '');
+        throw new Error(`NOTE_CREATE_FAILED:${noteResponse.status}:${detail}`);
     }
 
     const noteData = (await noteResponse.json().catch(() => ({}))) as HubSpotObjectResponse;
@@ -271,6 +276,7 @@ export async function createFallbackNote(token: string, contactId: string, body:
     }
 
     if (!associationResponse.ok) {
-        throw new Error(`NOTE_ASSOCIATION_FAILED:${associationResponse.status}`);
+        const detail = await associationResponse.text().catch(() => '');
+        throw new Error(`NOTE_ASSOCIATION_FAILED:${associationResponse.status}:${detail}`);
     }
 }
