+++ tests/chatbot-regressions.test.ts
@@ -1,7 +1,6 @@
 import test from 'node:test';
 import assert from 'node:assert/strict';
-import { detectBlockedDestination } from '../lib/ai/validation.ts';
-import { SYSTEM_INSTRUCTION } from '../lib/ai/prompt.ts';
+import { detectBlockedDestination, SYSTEM_INSTRUCTION } from '../api/generate.ts';
 
 test('safety check should block "Emirados" (Policy #186)', () => {
     const result = detectBlockedDestination('Dubai, Emirados Árabes Unidos');
